Refactored unescape. It is now 11 times faster than the previous one.
Bumped Net7 up to Net8, because UTF32.TryGetChars is only available in Net8.
Now there are only two bottlenecks:
```csharp
var result = token.Text[1..^1]; // allocates new sliced string
```
and
```csharp
result = new ReadOnlySpan<char>(p, result.Length - eaten).ToString(); // alocate x2 ><
```
Benchmark results, input string `string s = "\"\\'hello, world\\\"\\?\\\\test\\b\\f\\n\\r42\\t\\v \\02 \\024 \\007 \\077 \\x06 \\x28 \\u2200 \\U0001f34c \"";`:
```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22000.2538/21H2/SunValley)
Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
```
| Method | Mean       | Error    | StdDev   | Ratio         | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------- |-----------:|---------:|---------:|--------------:|--------:|-------:|----------:|------------:|
| New |   545.2 ns |  2.86 ns |  2.39 ns | 11.32x faster |   0.13x | 0.1450 |     608 B |  6.68x less |
| Old | 6,175.4 ns | 78.12 ns | 73.08 ns |      baseline |         | 0.9689 |    4064 B |             |